+++ packages/dashboard/src/components/Card/Card.types.ts
@@ -122,6 +122,12 @@ export interface ICardProps {
    * One of its use could be to fetch server data here
    */
   callOnDidMount?: VoidFunction;
+
+  /**
+   * Whether the card is draggable or not
+   * @default false
+   */
+  disableDrag?: boolean;
 }
 
 export interface ICardState {