+++ packages/example-app-base/src/components/templates/PageContent.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { BaseComponent, IBaseProps, IClassNames, customizable } from 'office-ui-fabric-react/lib/Utilities';
-import { ITheme, IStyle, mergeStyleSets } from 'office-ui-fabric-react/lib/Styling';
+import { BaseComponent, IBaseProps, customizable } from 'office-ui-fabric-react/lib/Utilities';
+import { ITheme, IStyle, mergeStyleSets, IProcessedStyleSet } from 'office-ui-fabric-react/lib/Styling';
 
 export interface IPageContentProps extends React.Props<PageContent>, IBaseProps {
   theme?: ITheme;
@@ -20,7 +20,7 @@ const getDefaultStyles = (props: IPageContentProps): IPageContentStyles => ({
 export class PageContent extends BaseComponent<IPageContentProps, {}> {
   public render(): JSX.Element {
     const { children } = this.props;
-    const classNames: IClassNames<IPageContentStyles> = mergeStyleSets(getDefaultStyles(this.props));
+    const classNames: IProcessedStyleSet<IPageContentStyles> = mergeStyleSets(getDefaultStyles(this.props));
 
     return <div className={classNames.root}>{children}</div>;
   }