+++ packages/office-ui-fabric-react/src/components/Callout/CalloutContent.base.tsx
@@ -140,6 +140,7 @@ export class CalloutContentBase extends BaseComponent<ICalloutProps, ICalloutSta
     let { target } = this.props;
     const {
       styles,
+      style,
       role,
       ariaLabel,
       ariaDescribedBy,
@@ -151,6 +152,7 @@ export class CalloutContentBase extends BaseComponent<ICalloutProps, ICalloutSta
       calloutWidth,
       calloutMaxWidth,
       finalHeight,
+      hideOverflow = !!finalHeight,
       backgroundColor,
       calloutMaxHeight,
       onScroll
@@ -165,7 +167,7 @@ export class CalloutContentBase extends BaseComponent<ICalloutProps, ICalloutSta
       calloutMaxHeight! && getContentMaxHeight && calloutMaxHeight! < getContentMaxHeight
         ? calloutMaxHeight!
         : getContentMaxHeight!;
-    const overflowYHidden = !!finalHeight;
+    const overflowYHidden = hideOverflow;
 
     const beakVisible = isBeakVisible && !!target;
     this._classNames = getClassNames(styles!, {
@@ -179,9 +181,12 @@ export class CalloutContentBase extends BaseComponent<ICalloutProps, ICalloutSta
       calloutMaxWidth
     });
 
-    const overflowStyle: React.CSSProperties = overflowYHidden
-      ? { overflowY: 'hidden', maxHeight: contentMaxHeight }
-      : { maxHeight: contentMaxHeight };
+    const overflowStyle: React.CSSProperties = {
+      ...style,
+      maxHeight: contentMaxHeight,
+      ...(overflowYHidden && { overflowY: 'hidden' })
+    };
+
     const visibilityStyle: React.CSSProperties | undefined = this.props.hidden ? { visibility: 'hidden' } : undefined;
     // React.CSSProperties does not understand IRawStyle, so the inline animations will need to be cast as any for now.
     const content = (