+++ packages/office-ui-fabric-react/src/components/Check/Check.base.tsx
@@ -3,7 +3,10 @@ import { BaseComponent } from '../../Utilities';
 import { ICheckProps } from './Check.types';
 import { Icon } from '../../Icon';
 import { getStyles } from './Check.styles';
-import { getClassNames } from './Check.classNames';
+import { classNamesFunction } from '../../Utilities';
+import { ICheckStyleProps, ICheckStyles } from './Check.types';
+
+const getClassNames = classNamesFunction<ICheckStyleProps, ICheckStyles>();
 
 export class CheckBase extends BaseComponent<ICheckProps, {}> {
   public static defaultProps: ICheckProps = {