+++ packages/office-ui-fabric-react/src/components/Coachmark/Coachmark.types.ts
@@ -1,11 +1,15 @@
 import { ICoachmarkStyles, ICoachmarkStyleProps } from './Coachmark.styles';
 import { IPositioningContainerTypes } from './PositioningContainer/PositioningContainer.types';
-import { IBaseProps, IRefObject, IStyleFunctionOrObject } from '../../Utilities';
+import { IRefObject, IStyleFunctionOrObject } from '../../Utilities';
+import { Coachmark } from './Coachmark';
 import { ITeachingBubble } from '../../TeachingBubble';
 
 export interface ICoachmark {}
 
-export interface ICoachmarkTypes extends IBaseProps<ICoachmark> {
+/** @deprecated */
+export type ICoachmarkTypes = ICoachmarkProps;
+
+export interface ICoachmarkProps extends React.Props<Coachmark> {
   /**
    * Optional callback to access the ICoachmark interface. Use this instead of ref for accessing
    * the public methods and properties of the component.