+++ packages/office-ui-fabric-react/src/components/DetailsList/examples/DetailsList.Grouped.Example.tsx
@@ -3,7 +3,7 @@ import { BaseComponent, createRef } from 'office-ui-fabric-react/lib/Utilities';
 import { DefaultButton } from 'office-ui-fabric-react/lib/Button';
 import { Fabric } from 'office-ui-fabric-react/lib/Fabric';
 import { IDetailsList, DetailsList, IColumn } from 'office-ui-fabric-react/lib/DetailsList';
-import { Checkbox } from '../../..';
+import { Checkbox } from 'office-ui-fabric-react/lib/Checkbox';
 import './DetailsList.Grouped.Example.scss';
 
 const _columns = [
@@ -119,6 +119,7 @@ export class DetailsListGroupedExample extends BaseComponent<
             showEmptyGroups: true
           }}
           onRenderItemColumn={this._onRenderColumn}
+          useReducedRowRenderer={true}
         />
       </Fabric>
     );