+++ packages/dashboard/CHANGELOG.md
@@ -1,6 +1,53 @@
 # Change Log - @uifabric/dashboard
 
-This log was last generated on Tue, 17 Jul 2018 10:28:40 GMT and should not be manually modified.
+This log was last generated on Wed, 25 Jul 2018 03:20:34 GMT and should not be manually modified.
+
+## 0.7.0
+Wed, 25 Jul 2018 03:20:34 GMT
+
+### Minor changes
+
+- Fix dashboard package webpack issue
+
+## 0.6.1
+Tue, 24 Jul 2018 10:24:36 GMT
+
+*Version update only*
+
+## 0.6.0
+Mon, 23 Jul 2018 10:28:08 GMT
+
+### Minor changes
+
+- Added disable drag prop to card to allow recommendation card to be non draggable
+- Addressing bad imports.
+
+## 0.5.0
+Fri, 20 Jul 2018 10:25:21 GMT
+
+### Minor changes
+
+- Brought in drag api for dashboard grid layout, which enables adding elements from outside the grid into the grid layout
+
+## 0.4.5
+Thu, 19 Jul 2018 21:25:32 GMT
+
+*Version update only*
+
+## 0.4.4
+Thu, 19 Jul 2018 19:04:38 GMT
+
+*Version update only*
+
+## 0.4.3
+Thu, 19 Jul 2018 10:23:34 GMT
+
+*Version update only*
+
+## 0.4.2
+Wed, 18 Jul 2018 10:25:50 GMT
+
+*Version update only*
 
 ## 0.4.1
 Tue, 17 Jul 2018 10:28:40 GMT