+++ packages/utilities/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @uifabric/utilities
 
-This log was last generated on Tue, 17 Jul 2018 10:28:40 GMT and should not be manually modified.
+This log was last generated on Mon, 23 Jul 2018 10:28:08 GMT and should not be manually modified.
+
+## 6.12.0
+Mon, 23 Jul 2018 10:28:08 GMT
+
+### Minor changes
+
+- Addressing bad imports.
+
+## 6.11.0
+Wed, 18 Jul 2018 10:25:50 GMT
+
+### Minor changes
+
+- Tighten typings of APIs around merge-styles to be more correct
 
 ## 6.10.0
 Tue, 17 Jul 2018 10:28:40 GMT